Seperated Components for ContractPage
