+++ .gitignore
@@ -8,3 +8,4 @@ report.json
 dist
 .flyway
 **/cypress/screenshots/
+services/client/src/CrownsScratch.md
